Add files via upload
The Case Study
Analyzing an affiliated credit card (special card only available to members of a specific group). Think of it like this the credit card company is like Citibank and the affiliated group is a company like Indian Oil or Barista. The affiliated group has to pay some fees to credit card company (because the credit card company is providing this service to the group) Product Details :
-	average card balance is Rs 1000
-	each card has an interest rate of 15%
-	membership fee is Rs 20 per card
-	loss rate is 3% (NOTE: what are the interpretations of this ? state your assumption CLEARLY when solving  )
-	Rs 25 operating cost per card
-	Rs 10 affiliation fee per card (cost to the group organization itself, so that their members get this awesome card program)
-	6.5% cost of funds (that credit card company must pay to the bank where it gets money)
-	bill/statement is issued on 1st of every month
-	credit-free period = 15 days from the bill/statement issue date.
Provide a monthly calculation of profits/losses given :
i) The interest rate is simple interest. ii) Interest is calculated on a daily basis.
iii) The average card balance is the average end of month balance. iv) All members are more than a year old.
Read more at: Link
There might be some terminology that may not be familiar to you. This is expected. We want to see if you can research and find out about them and then answer these questions.
Consider 1000 customers for the above scenario, divide these 1000 customers into following groups(upon your assumption and understanding about the credit lending industry)
1.customers who pay the bill within the credit-free period
2.customers paying 30 days after the last due date
3.customer paying 45 days after the last due date
4.customer paying 60 days after the last due date(remove these customers for the complete calculation of the next month,basically, you are not supposed to provide further credit services to these customers)
Answer the following for both scenarios.
1.	How much profit % is generated in 1 year?
2.	The 3% loss rate can either mean either 3% of the outstanding balance defaults or 3% of the borrowers default on their entire balance. If the average balance is increased to $2000, what happens to the loss?
3.	Is a borrower with a low balance more or less likely to default then a borrower with a high balance? Why ? explain clearly
4.	The group company has now become unhappy with the credit card provider. It wants to start its own credit card. But for that, it will need to purchase the customers from the current credit card company. If the affiliated group offered Rs 20 per account to purchase all the customers from the credit card company, what would their profit be? (assume that we are only talking about 1 year)
5.	What numbers can the financial institution change to convince the affiliated group not to want to purchase, while still not going into a loss? Give a range of what the new numbers are (min and max)
6.	What is the extra cost(if any) company will have to incur if the customer delays the payment of the credit card bill after the due date?
7.	Is it beneficial for the company if the customer pays the credit card bill after the due date? (express your views as a business strategy analyst with a short example)